Rename pyomo.common.testing so nose doesn't think it is a test
